requirements.txt
django==1.11
djangorestframework
psycopg2-binary
redis
requests
